Basic snap shotting seems OK

+++ shared-binary/include/g_industrial_cam/device-io/camera.hpp
@@ -13,7 +13,7 @@ namespace g_industrial_cam{
         camera() = delete;
         camera(const std::string& identifier_utf8);
         ~camera();
-        buffer* take_snapshot(uint32_t timeout) const;
+        buffer* take_snapshot(uint64_t timeout) const;
         void get_avaliable_pixel_formats(std::vector<std::string>& pixel_formats_utf8) const;
         std::string get_pixel_format() const;
         void set_pixel_format(const std::string& pixel_format_utf8);
